requirements.txt
black
opencv-python
python-dotenv
pillow
pillow-heif
exifread
defusedxml
chromadb
pymongo
ollama
pandas
streamlit